Applied r3328, r3329, r3332, r3333, r3334, r3339, r3343, r3344, r3345, r3346, r3347, r3348, r3350, r3352, r3353, r3354, r3355, r3356, r3357, r3358, r3361, r3365, r3366, r3367, r3368, r3369, r3380, r3381, r3382, r3385, r3389, r3390, r3395, r3396, r3401, r3422, r3423, r3439, r3441r3328, r3329, r3332, r3333, r3334, r3339, r3343, r3344, r3345, r3346, r3347, r3348, r3350, r3352, r3353, r3354, r3355, r3356, r3357, r3358, r3361, r3365, r3366, r3367, r3368, r3369, r3380, r3381, r3382, r3385, r3389, r3390, r3395, r3396, r3401, r3422, r3423, r3439, r3441 from sogo-good branch

+++ utils/mapitest/mapitest_stat.c
@@ -98,7 +98,7 @@ _PUBLIC_ uint32_t mapitest_stat_add_result(struct mapitest_suite *suite,
 		if (testresult == Fail) {
 			suite->stat->failure++;
 			el->reason = talloc_strdup((TALLOC_CTX *) el, "Unexpected Fail");
-		} else if (testresult == ExpectedFail) {
+		} else if ((int)testresult == (int)ExpectedFail) {
 			suite->stat->x_fail++;
 			el->reason = talloc_strdup((TALLOC_CTX *) el, "Expected Fail");
 		} else if (testresult == UnexpectedPass) {
